Get extension with authtoken and apikey code added to manage fallback
